abstract handleAddPairFilter/handleRemovePairFilter to utils
