+++ src/state/utils.js
@@ -4,7 +4,7 @@ import _omit from 'lodash/omit'
 
 import { platform } from 'var/config'
 import { getPath, TYPE_WHITELIST, ROUTE_WHITELIST } from 'state/query/utils'
-import { getSymbolsURL } from 'state/symbols/utils'
+import { getSymbolsURL, parsePairTag } from 'state/symbols/utils'
 
 export function postJsonfetch(url, bodyJson) {
   return fetch(url, {
@@ -97,6 +97,14 @@ export function handleAddSymbolFilter(type, symbol, props) {
   }
 }
 
+export function handleAddPairFilter(type, pair, props) {
+  const { history, addTargetPair, targetPairs } = props
+  if (!targetPairs.includes(pair)) {
+    history.push(generateUrl(type, history.location.search, [...targetPairs, pair]))
+    addTargetPair(pair)
+  }
+}
+
 export function handleRemoveSymbolFilter(type, tag, props) {
   const { history, removeTargetSymbol, targetSymbols } = props
   if (targetSymbols.includes(tag)) {
@@ -106,6 +114,19 @@ export function handleRemoveSymbolFilter(type, tag, props) {
       history.push(generateUrl(type, history.location.search, targetSymbols.filter(symbol => symbol !== tag)))
     }
     removeTargetSymbol(tag)
+  }
+}
+
+export function handleRemovePairFilter(type, tag, props) {
+  const { history, removeTargetPair, targetPairs } = props
+  const parsedTag = parsePairTag(tag)
+  if (targetPairs.includes(parsedTag)) {
+    if (targetPairs.length === 1) { // show no select symbol in url
+      history.push(generateUrl(type, history.location.search))
+    } else {
+      history.push(generateUrl(type, history.location.search, targetPairs.filter(pair => pair !== parsedTag)))
+    }
+    removeTargetPair(parsedTag)
   }
 }
 
@@ -161,7 +182,9 @@ export default {
   getParsedUrlParams,
   getSideMsg,
   generateUrl,
+  handleAddPairFilter,
   handleAddSymbolFilter,
+  handleRemovePairFilter,
   handleRemoveSymbolFilter,
   isValidateType,
   momentFormatter,
